refactor: replace raw Error throws with domain exceptions in auth use cases
- Replace all 'throw new Error(...)' with appropriate domain exceptions
- Use ValidationException for user not found scenarios
- Use AuthenticationException for token/verification failures
- Use ConflictException for already verified users
- Ensure consistent error handling across all auth use cases
- Remove ad-hoc error handling patterns
