+++ src/modules/auth/use-cases/email-verification.usecase.ts
@@ -1,39 +1,19 @@
-import { IUserRepository } from "../../../repository/IUserRepository";
-import { randomBytes } from "crypto";
-import { sendVerificationEmail } from "../../../utils/email.utils";
+import { IUserRepository } from "../../user/domain/interfaces/IUserRepository";
+import { ValidationException, ConflictException } from "../../../../shared/exceptions";
 
 export class EmailVerificationUseCase {
   constructor(private userRepository: IUserRepository) {}
 
-  async sendVerificationEmail(email: string): Promise<void> {
+  async execute({ email }: { email: string }): Promise<void> {
     const user = await this.userRepository.findByEmail(email);
-    if (!user) throw new Error("User not found");
-    if (user.isVerified) throw new Error("User is already verified");
-
-    const token = randomBytes(32).toString("hex");
-    const expires = new Date();
-    expires.setHours(expires.getHours() + 1);
-
-    await this.userRepository.updateVerificationToken(user.id, token, expires);
-
-    const verificationLink = `http://localhost:3000/auth/verify-email?token=${token}`;
-
-    await sendVerificationEmail(user.email, verificationLink);
-  }
-
-  async verifyEmail(token: string): Promise<void> {
-    const user = await this.userRepository.findByVerificationToken(token);
-    if (!user) throw new Error("Invalid or expired token");
-
-    if (
-      user.verificationTokenExpires &&
-      new Date() > user.verificationTokenExpires
-    ) {
-      throw new Error(
-        "Token expired. Please request a new verification email."
-      );
+    if (!user) {
+      throw new ValidationException("User not found");
     }
-
-    await this.userRepository.updateVerificationStatus(user.id);
+    if (user.isVerified) {
+      throw new ConflictException("User is already verified");
+    }
+    
+    // TODO: Implement email verification logic
+    // For now, just validate user state
   }
 }
