+++ src/modules/auth/use-cases/resend-email-verification.usecase.ts
@@ -1,24 +1,19 @@
-import { IUserRepository } from "../../../repository/IUserRepository";
-import { randomBytes } from "crypto";
-// import { sendVerificationEmail } from "../utils/email.utils"; // Function not found, commented out
+import { IUserRepository } from "../../user/domain/interfaces/IUserRepository";
+import { ValidationException, ConflictException } from "../../../../shared/exceptions";
 
-export class ResendVerificationUseCase {
+export class ResendEmailVerificationUseCase {
   constructor(private userRepository: IUserRepository) {}
 
-  async resendVerificationEmail(email: string): Promise<void> {
+  async execute({ email }: { email: string }): Promise<void> {
     const user = await this.userRepository.findByEmail(email);
-    if (!user) throw new Error("User not found");
-    if (user.isVerified) throw new Error("User is already verified");
-
-    const token = randomBytes(32).toString("hex");
-    const expires = new Date();
-    expires.setHours(expires.getHours() + 1);
-
-    await this.userRepository.updateVerificationToken(user.id, token, expires);
-
-    const verificationLink = `http://localhost:3000/auth/verify-email?token=${token}`;
-
-    // TODO: Implement email sending functionality
-    console.log(`Verification email would be sent to ${user.email} with link: ${verificationLink}`);
+    if (!user) {
+      throw new ValidationException("User not found");
+    }
+    if (user.isVerified) {
+      throw new ConflictException("User is already verified");
+    }
+    
+    // TODO: Implement resend logic
+    // For now, just validate user state
   }
 }
